print

+++ onmt/Trainer.py
@@ -42,7 +42,7 @@ class Statistics(object):
         self.n_correct += stat.n_correct
 
     def accuracy(self):
-        return 100 * (self.n_correct / self.n_words.float())    # modify
+        return 100 * (self.n_correct / self.n_words)    # modify
 
     def xent(self):
         return self.loss / self.n_words.float()     # modify
